+++ pytsite/widget/_static.py
@@ -48,45 +48,6 @@ class Text(_base.Abstract):
         return self._group_wrap(container)
 
 
-class Tabs(_base.Abstract):
-    """Tabs Widget.
-    """
-
-    def __init__(self, uid: str, **kwargs):
-        """Init.
-        """
-        super().__init__(uid, **kwargs)
-        self._tabs = []
-
-    def add_tab(self, tid: str, title: str, content: str):
-        """Add a tab.
-        """
-        tid = tid.replace('.', '-')
-        self._tabs.append((tid, title, content))
-        return self
-
-    def get_html_em(self, **kwargs) -> str:
-        wrapper = _html.Div(role='tabpanel')
-        tabs_ul = _html.Ul(cls='nav nav-tabs', role='tablist')
-        content = _html.Div(cls='tab-content')
-        wrapper.append(tabs_ul).append(content)
-
-        i = 0
-        for tab in self._tabs:
-            tab_uid = 'tab-uid-' + tab[0]
-            tabs_ul.append(
-                _html.Li(role='presentation', cls='active' if i == 0 else '').append(
-                    _html.A(tab[1], href='#' + tab_uid, role='tab', data_toggle='tab')
-                )
-            )
-            content_cls = 'tabpanel tab-pane'
-            content_cls += ' active' if i == 0 else ''
-            content.append(_html.Div(tab[2], cls=content_cls, uid=tab_uid))
-            i += 1
-
-        return self._group_wrap(wrapper)
-
-
 class VideoPlayer(_base.Abstract):
     """Video player widget.
     """